FILE: moves create-dtos functions and mock-data file to data-access
