+++ libs/menu-matriarch/data-access/src/lib/internal/dish-data.service.ts
@@ -4,12 +4,8 @@ import { map } from 'rxjs/operators';
 
 import { DataService } from '@atocha/core/data-access';
 import { lower, sort } from '@atocha/core/util';
-import {
-  Dish,
-  DishDto,
-  Endpoint,
-  createDishDto,
-} from '@atocha/menu-matriarch/util';
+import { Dish, DishDto, Endpoint } from '@atocha/menu-matriarch/util';
+import { createDishDto } from '../dtos/create-dtos';
 import { BatchService } from './batch.service';
 
 export type EditableDishData = Pick<
