+++ libs/menu-matriarch/util/src/index.ts
@@ -1,8 +1,5 @@
-export * from './lib/functions/create-dtos';
 export * from './lib/functions/get-days';
 export * from './lib/functions/get-dish-types';
-
-export * from './lib/mappers/mock-data';
 
 export * from './lib/day.type';
 export * from './lib/dish.interface';
